+++ test/test_helper.exs
@@ -1,3 +1,4 @@
 ExUnit.start()
 
 JOSE.crypto_fallback(true)
+JOSE.unsecured_signing(true)