+++ test/validator/jwt_test.exs
@@ -22,6 +22,8 @@ defmodule APIacAuthBearerTest.Validator.JWT do
     }
   }
 
+  @default_jwt_header %{"typ" => "at+jwt"}
+
   @valid_jwt_payload %{
     "iss" => "some issuer",
     "exp" => System.system_time(:second) + 30,
@@ -53,95 +55,101 @@ defmodule APIacAuthBearerTest.Validator.JWT do
   end
 
   describe ".validate_bearer/2" do
-    test "valid signed AT with RSA signature" do
+    test "valid signed AT with RSA signature is accepted" do
       bearer = @valid_jwt_payload |> gen_jwt("RS256")
 
       assert {:ok, @valid_jwt_payload} = JWT.validate_bearer(bearer, @valid_options)
     end
 
-    test "valid signed AT with EC signature" do
+    test "valid signed AT with EC signature is accepted" do
       bearer = @valid_jwt_payload |> gen_jwt("ES256")
 
       assert {:ok, @valid_jwt_payload} = JWT.validate_bearer(bearer, @valid_options)
     end
 
-    test "valid signed AT with EdDSA signature" do
+    test "valid signed AT with EdDSA signature is accepted" do
       bearer = @valid_jwt_payload |> gen_jwt("EdDSA")
 
       assert {:ok, @valid_jwt_payload} = JWT.validate_bearer(bearer, @valid_options)
     end
 
-    test "valid signed AT with HS256 MACing" do
+    test "valid signed AT with HS256 MACing is accepted" do
       bearer = @valid_jwt_payload |> gen_jwt("HS256")
 
       assert {:ok, @valid_jwt_payload} = JWT.validate_bearer(bearer, @valid_options)
     end
 
-    test "signed with invalid typ" do
+    test "unsigned AT is rejected" do
+      bearer = @valid_jwt_payload |> gen_jwt("none")
+
+      assert {:error, _} = JWT.validate_bearer(bearer, @valid_options)
+    end
+
+    test "signed with invalid typ is rejected" do
       bearer = @valid_jwt_payload |> gen_jwt("ES256", %{"typ" => "invalid+jwt"})
 
       assert {:error, _} = JWT.validate_bearer(bearer, @valid_options)
     end
 
-    test "signed with missing typ" do
-      bearer = @valid_jwt_payload |> gen_jwt("ES256", %{})
+    test "signed with nill typ is rejected" do
+      bearer = @valid_jwt_payload |> gen_jwt("ES256", %{"typ" => nil})
 
       assert {:error, _} = JWT.validate_bearer(bearer, @valid_options)
     end
 
-    test "signed with invalid issuer" do
+    test "signed with invalid issuer is rejected" do
       bearer = @valid_jwt_payload |> Map.put("iss", "invalid") |> gen_jwt("ES256")
 
       assert {:error, _} = JWT.validate_bearer(bearer, @valid_options)
     end
 
-    test "signed with expired JWT" do
+    test "signed with expired JWT is rejected" do
       bearer = @valid_jwt_payload |> Map.put("exp", now() - 1) |> gen_jwt("ES256")
 
       assert {:error, _} = JWT.validate_bearer(bearer, @valid_options)
     end
 
-    test "signed with invalid signature" do
+    test "signed with invalid signature is rejected" do
       bearer = @valid_jwt_payload |> gen_jwt("ES256") |> Kernel.<>("z")
 
       assert {:error, _} = JWT.validate_bearer(bearer, @valid_options)
     end
 
-    test "encrypted AT with missing client config" do
+    test "encrypted AT with missing client config is rejected" do
       bearer = @valid_jwt_payload |> gen_jwt("RS256", "RSA-OAEP", "A128GCM")
 
       assert {:error, _} = JWT.validate_bearer(bearer, @valid_options)
     end
 
-    test "valid encrypted then signed AT with RSA encryption" do
+    test "valid encrypted then signed AT with RSA encryption is accepted" do
       bearer = @valid_jwt_payload |> gen_jwt("RS256", "RSA-OAEP", "A128GCM")
       opts = Keyword.put(@valid_options, :client_config, &client_config_enc_rsa_oaep/1)
 
       assert {:ok, @valid_jwt_payload} = JWT.validate_bearer(bearer, opts)
     end
 
-    test "valid encrypted then signed AT with ECDH-ES encryption" do
+    test "valid encrypted then signed AT with ECDH-ES encryption is accepted" do
       bearer = @valid_jwt_payload |> gen_jwt("RS256", "ECDH-ES", "A128GCM")
       opts = Keyword.put(@valid_options, :client_config, &client_config_enc_ecdh_es/1)
 
       assert {:ok, @valid_jwt_payload} = JWT.validate_bearer(bearer, opts)
     end
 
-    test "valid encrypted then signed AT with dir symmetric encryption" do
+    test "valid encrypted then signed AT with dir symmetric encryption is accepted" do
       bearer = @valid_jwt_payload |> gen_jwt("RS256", "dir", "A256GCM")
       opts = Keyword.put(@valid_options, :client_config, &client_config_enc_dir/1)
 
       assert {:ok, @valid_jwt_payload} = JWT.validate_bearer(bearer, opts)
     end
 
-    test "encrypted then signed AT with invalid at_encrypted_response_alg value" do
+    test "encrypted then signed AT with invalid at_encrypted_response_alg value is rejected" do
       bearer = @valid_jwt_payload |> gen_jwt("RS256", "RSA-OAEP", "A256GCM")
       opts = Keyword.put(@valid_options, :client_config, &client_config_enc_dir/1)
 
       assert {:error, _} = JWT.validate_bearer(bearer, opts)
     end
 
-    test "encrypted then signed AT with invalid type" do
+    test "encrypted then signed AT with invalid type is rejected" do
       bearer =
         @valid_jwt_payload |> gen_jwt("RS256", "RSA-OAEP", "A128GCM", %{"typ" => "x("})
       opts = Keyword.put(@valid_options, :client_config, &client_config_enc_rsa_oaep/1)
@@ -149,14 +157,14 @@ defmodule APIacAuthBearerTest.Validator.JWT do
       assert {:error, _} = JWT.validate_bearer(bearer, opts)
     end
 
-    test "encrypted then signed AT with missing type" do
-      bearer = @valid_jwt_payload |> gen_jwt("RS256", "RSA-OAEP", "A128GCM", %{})
+    test "encrypted then signed AT with nil type is rejected" do
+      bearer = @valid_jwt_payload |> gen_jwt("RS256", "RSA-OAEP", "A128GCM", %{"typ" => nil})
       opts = Keyword.put(@valid_options, :client_config, &client_config_enc_rsa_oaep/1)
 
       assert {:error, _} = JWT.validate_bearer(bearer, opts)
     end
 
-    test "encrypted then signed AT with invalid encryption" do
+    test "encrypted then signed AT with invalid encryption is rejected" do
       bearer = @valid_jwt_payload |> gen_jwt("RS256", "RSA-OAEP", "A128GCM")
       [header, encrypted_key, iv, ciphertext, tag] = String.split(bearer, ".")
       bearer = Enum.join([header, encrypted_key, iv, ciphertext <> "invalid", tag])
@@ -166,7 +174,15 @@ defmodule APIacAuthBearerTest.Validator.JWT do
     end
   end
 
-  defp gen_jwt(payload, sig_alg, additional_headers \\ %{"typ" => "at+jwt"}) do
+  defp gen_jwt(payload, sig_alg, additional_headers \\ %{})
+
+  defp gen_jwt(payload, "none", additional_headers) do
+    payload
+    |> Jason.encode!()
+    |> JOSEUtils.JWS.sign!(%{}, "none", Map.merge(@default_jwt_header, additional_headers))
+  end
+
+  defp gen_jwt(payload, sig_alg, additional_headers) do
     [signing_key | _] =
       @server_metadata["jwks"]["keys"]
       |> JOSEUtils.JWKS.signature_keys()
@@ -174,11 +190,11 @@ defmodule APIacAuthBearerTest.Validator.JWT do
 
     payload
     |> Jason.encode!()
-    |> JOSEUtils.JWS.sign!(signing_key, sig_alg, additional_headers)
+    |> JOSEUtils.JWS.sign!(signing_key, sig_alg, Map.merge(@default_jwt_header, additional_headers))
   end
 
   defp gen_jwt(
-    payload, sig_alg, enc_alg, enc_enc, additional_headers \\ %{"typ" => "at+jwt"}
+    payload, sig_alg, enc_alg, enc_enc, additional_headers \\ %{}
   ) do
     jws = gen_jwt(payload, sig_alg, %{})
 
@@ -199,7 +215,7 @@ defmodule APIacAuthBearerTest.Validator.JWT do
         encryption_key
       end
 
-    JOSEUtils.JWE.encrypt!(jws, jwk, enc_alg, enc_enc, additional_headers)
+    JOSEUtils.JWE.encrypt!(jws, jwk, enc_alg, enc_enc, Map.merge(@default_jwt_header, additional_headers))
   end
 
   def client_config_enc_rsa_oaep(_opts),