+++ CHANGELOG.md
@@ -4,13 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog and this project adheres to Semantic Versioning.
 
-## [2.0.0] - FIXME
+## [2.0.0] - 2021-11-05
 
 ### Added
 
-- [`APIacAuthBearer.Validator.JWT`]: new JWT validator
+- [`APIacAuthBearer.Validator.JWT`]: add RFC9068 JWT validator
 
 ### Changed
 
 - [`APIacAuthBearer`]: `:resource_server_name` option becomes `:resource_indicator` to
-match the specifications better
+better match the specifications